app/requirements.txt
fastapi
uvicorn[standard]
motor
pillow
numpy
faiss-cpu
insightface
python-multipart
social-analyzer
